Pulling out common http-based test logic
